Fixed up plumbing issues so BrowserTests could be run
