+++ GlyssenTests/Controls/BrowserTests.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Windows.Forms;
 using Glyssen.Controls;
+using Glyssen.Utilities;
 using NUnit.Framework;
 
 namespace GlyssenTests.Controls
 {
-	class BrowserTests
+	[TestFixture]
+	public class BrowserTests
 	{
 		[Test, Ignore("By hand only")]
 		[STAThread]
@@ -13,6 +15,8 @@ namespace GlyssenTests.Controls
 		{
 			using (var form = new Form())
 			{
+				GeckoUtilities.InitializeGecko();
+
 				using (var browser = new Browser())
 				{
 					form.Controls.Add(browser);
@@ -28,6 +32,8 @@ namespace GlyssenTests.Controls
 		{
 			using (var form = new Form())
 			{
+				GeckoUtilities.InitializeGecko();
+
 				using (var browser = new Browser())
 				{
 					form.Controls.Add(browser);
