+++ GlyssenTests/packages.config
@@ -6,6 +6,7 @@
   <package id="DotNetZip" version="1.13.4" targetFramework="net472" />
   <package id="Enchant.Net" version="1.4.2" targetFramework="net472" />
   <package id="GdkSharp" version="3.22.24.37" targetFramework="net472" />
+  <package id="Geckofx60.64" version="60.0.44" targetFramework="net472" />
   <package id="GioSharp" version="3.22.24.37" targetFramework="net472" />
   <package id="GLibSharp" version="3.22.24.37" targetFramework="net472" />
   <package id="GtkSharp" version="3.22.24.37" targetFramework="net472" />
@@ -38,6 +39,7 @@
   <package id="SIL.TestUtilities" version="8.0.0-beta0037" targetFramework="net472" />
   <package id="SIL.Windows.Forms" version="8.0.0-beta0037" targetFramework="net472" />
   <package id="SIL.Windows.Forms.DblBundle" version="8.0.0-beta0033" targetFramework="net472" />
+  <package id="SIL.Windows.Forms.GeckoBrowserAdapter" version="8.0.0-beta0033" targetFramework="net472" />
   <package id="SIL.Windows.Forms.Keyboarding" version="8.0.0-beta0033" targetFramework="net472" />
   <package id="SIL.Windows.Forms.Scripture" version="8.0.0-beta0037" targetFramework="net472" />
   <package id="SIL.Windows.Forms.WritingSystems" version="7.0.0" targetFramework="net472" />
